chore: bump toolchain and mpc versions
This PR bumps toolchain to v0.7.0-3-g45b5b5d and mpc to 1.3.0. Renovate bot is also showing updates for golang, but it's an RC and not an actual release.

Finally, it's showing an update to protobuf, but their release process seems to not have published tar.gz files this time. I've added a thread to their forum to ask for clarity about this:

https://groups.google.com/g/protobuf/c/I4EXb4PhYyc

Signed-off-by: Spencer Smith <spencer.smith@talos-systems.com>